refactor(#13): Refactor folder naming in the project
